[DEV-2269] refactoring
