+++ src/subdomains/supporting/payment/dto/transaction.dto.ts
@@ -1,0 +1,6 @@
+export enum TransactionError {
+  AMOUNT_TOO_LOW = 'AmountTooLow',
+  AMOUNT_TOO_HIGH = 'AmountTooHigh',
+  BANK_TRANSACTION_MISSING = 'BankTransactionMissing',
+  KYC_REQUIRED = 'KycRequired',
+}
