+++ src/subdomains/supporting/payment/services/transaction-helper.ts
@@ -16,6 +16,7 @@ import { PricingService } from '../../pricing/services/pricing.service';
 import { FeeDto } from '../dto/fee.dto';
 import { FiatPaymentMethod, PaymentMethod } from '../dto/payment-method.enum';
 import { TargetEstimation, TransactionDetails } from '../dto/transaction-details.dto';
+import { TransactionError } from '../dto/transaction.dto';
 import { TxFeeDetails } from '../dto/tx-fee-details.dto';
 import { TxSpec, TxSpecExtended } from '../dto/tx-spec.dto';
 import { TransactionDirection, TransactionSpecification } from '../entities/transaction-specification.entity';
@@ -24,13 +25,6 @@ import { TransactionSpecificationRepository } from '../repositories/transaction-
 export enum ValidationError {
   PAY_IN_TOO_SMALL = 'PayInTooSmall',
   PAY_IN_NOT_SELLABLE = 'PayInNotSellable',
-}
-
-export enum TransactionError {
-  AMOUNT_TOO_LOW = 'AmountTooLow',
-  AMOUNT_TOO_HIGH = 'AmountTooHigh',
-  BANK_TRANSACTION_MISSING = 'BankTransactionMissing',
-  KYC_REQUIRED = 'KycRequired',
 }
 
 @Injectable()
